% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisStatistics.R
\name{AnalysisStatistics}
\alias{AnalysisStatistics}
\title{Count Statistics for ChIA-PET.}
\usage{
AnalysisStatistics(x.self, x.intra = NULL, x.inter = NULL,
  file.out = NULL, threshold = 1e-05, savedir = NULL)
}
\arguments{
\item{x.self}{An object of class \code{\linkS4class{PSelf}} or
\code{\linkS4class{PSFit}}.}

\item{x.intra}{An object of class \code{\linkS4class{PIntra}}
(optional).}

\item{x.inter}{An object of class \code{\linkS4class{PInter}}
(optional).}

\item{file.out}{A string with the name of the output to be saved to
\code{savedir}. If \code{NULL} the function will only print the output.}

\item{threshold}{A numeric indicating the FDR cut-off, used when
\code{class(x.self)=\linkS4class{PSFit}}. If NULL, no threshold is applied.}

\item{savedir}{A string with the directory to save the ouput file. If
\code{NULL} then the function will only print the output.}
}
\value{
Based on the inputs, \code{AnalysisStatistics} prints the total
 Self-ligated, Intra- and Inter-chromosomal PETs, as well as the total regions,
  total candidate peaks and total significant
peaks (if \code{threshold!=NULL} and
\code{class(x.self)=\linkS4class{PSFit}}). If \code{file.out} and
\code{savedir} are not \code{NULL} then it also saves the output to a csv
file in \code{savedir}.
}
\description{
\code{AnalysisStatistics} prints and saves count statistics
for the current inputs of the peak-calling analysis.
}
\examples{
#Create a temporary test forder, or anywhere you want:
savedir=file.path(tempdir(),'MACPETtest')
dir.create(savedir)#where you will save the results

#load Inter-chromosomal data:
load(system.file('extdata', 'pinterData.rda', package = 'MACPET'))
class(pinterData)

#load Intra-chromosomal data:
load(system.file('extdata', 'pintraData.rda', package = 'MACPET'))
class(pintraData)

#################################################################
#load Self-ligated data: (class=PSelf)
load(system.file('extdata', 'pselfData.rda', package = 'MACPET'))
class(pselfData)

#Print analysis:
AnalysisStatistics(x.self=pselfData,
                   x.intra=pintraData,
                   x.inter=pinterData,
                   file.out='AnalysisStats',
                   savedir=savedir)

#################################################################
#load Self-ligated data: (class=PSFit)
load(system.file('extdata', 'psfitData.rda', package = 'MACPET'))
class(psfitData)

#Print analysis:
AnalysisStatistics(x.self=psfitData,
                   x.intra=pintraData,
                   x.inter=pinterData,
                   file.out='AnalysisStats',
                   savedir=savedir,
                   threshold=1e-5)

#-----delete test directory:
unlink(savedir,recursive=TRUE)
}
\references{
Vardaxis I, Drabløs F, Rye M and Lindqvist BH (2018). \emph{Model-based Analysis for ChIA-PET (MACPET)}.
To be published.
}
\seealso{
\code{\linkS4class{PSelf}},
 \code{\linkS4class{PSFit}},\code{\linkS4class{PIntra}},
\code{\linkS4class{PInter}}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
