% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPeaks-methods.R
\name{exportPeaks}
\alias{exportPeaks}
\alias{exportPeaks.default}
\alias{exportPeaks.PSFit}
\title{Exports peaks to csv file}
\usage{
exportPeaks(object, ...)

\method{exportPeaks}{default}(object, ...)

\method{exportPeaks}{PSFit}(object, file.out, savedir, threshold = NULL, ...)
}
\arguments{
\item{object}{An object of \code{\linkS4class{PSFit}}  class.}

\item{...}{Further arguments to be passed to \code{exportPeaks} (not used).}

\item{file.out}{A string with the name of the output to be saved to
\code{savedir}.}

\item{savedir}{A string with the directory to save the output.}

\item{threshold}{A numeric indicating the FDR cut-off used for subseting
significant peaks. If NULL all the peaks are returned.}
}
\value{
For \code{\linkS4class{PSFit}} class: a csv file named after the
value of \code{file.out} with all the information about the peaks found by the
\code{\link{PeakCallerUlt}} function, plus comments which
 explain the column names.
}
\description{
\code{exportPeaks} is an S3 method for the  \linkS4class{PSFit}
class. It exports peak information to a csv file in a given directory.
}
\examples{

#Create a test forder on the desktop, or anywhere you want:
savedir=file.path(path.expand('~'),'Desktop')
dir.create(file.path(savedir,"MACPET.test"))
savedir=file.path(savedir,"MACPET.test")#where you will save
#the results

#load Self-ligated data: (class=PSFit)
load(system.file("extdata", "psfitData.rda", package = "MACPET"))
class(psfitData)
exportPeaks(object=psfitData,file.out="Peaks",threshold=1e-5,savedir=savedir)

#-----delete test directory:
unlink(savedir,recursive=TRUE)
}
\references{
Vardaxis I, Drabløs F, Rye M and Lindqvist BH (2018). \emph{Model-based Analysis for ChIA-PET (MACPET)}.
To be published.
}
\seealso{
\linkS4class{PSFit}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
