% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GenomeMap-class}
\alias{GenomeMap-class}
\alias{GenomeMap}
\title{GenomeMap S4 Class}
\description{
\code{\linkS4class{GenomeMap}} class in a S4 class which
 inherits from the
 \code{\link[InteractionSet:GInteractions-class]{GInteractions}} class and it contains
all the interactions found by MACPET at Stage 4. This class represents
the genome map constructed by the peaks which interact. Furthermore, in the
 \code{\link[S4Vectors:Annotated-class]{metadata}} field, it contains a
 \code{\link[S4Vectors:DataFrame-class]{DataFrame}} called \code{InteractionInfo}
 with the interaction information. More precisely, the
 \code{InteractionInfo} object contains the following columns:
 \describe{
 \item{\code{pvalue}}{ The p-value of the interaction.}
 \item{\code{FDR}}{ The FDR of the interaction.}
 \item{\code{Order}}{ The order the interaction was entered into the model.}
 \item{\code{TotalInterPETs}}{ The total interaction PETs between every two interacting peaks.}
 }
 Finally it also contains two metadata columns with the peak summits of the two peaks in each interaction.
}
\details{
\code{\linkS4class{GenomeMap}} class is created by
the \code{\link{MACPETUlt}}
function at Stage 4. Interactions with same \code{Order} are those who entered the model together because they
had the same \code{FDR} value. Subsets of significant interactions can be found using the
\code{\link{GetSignInteractions}} function.
}
\references{
Vardaxis I, Drabløs F, Rye M and Lindqvist BH (2018). \emph{MACPET: Complete pipeline for ChIA-PET}.
To be published.
}
\seealso{
\code{\link{plot}},
\code{\link{summary}},\code{\link{GetSignInteractions}},
\code{\link{MACPETUlt}}, \code{\link{GetSortestPath}}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
