% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSignInteractions-methods.R
\docType{methods}
\name{GetSignInteractions}
\alias{GetSignInteractions}
\alias{GetSignInteractions.default}
\alias{GetSignInteractions.GenomeMap}
\alias{GetSignInteractions,GenomeMap-method}
\alias{GetSignInteractions,GenomeMap,GetSignInteractions-method}
\title{Get the significant interactions of a \code{\linkS4class{GenomeMap}} object}
\usage{
GetSignInteractions(object, ...)

\method{GetSignInteractions}{default}(object, ...)

\method{GetSignInteractions}{GenomeMap}(object, threshold = NULL,
  ReturnedAs = "GInteractions", SaveTo = "", Prefix = "MACPET", ...)

\S4method{GetSignInteractions}{GenomeMap}(object, threshold = NULL,
  ReturnedAs = "GInteractions", SaveTo = "", Prefix = "MACPET", ...)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{GenomeMap}}.}

\item{...}{Further arguments to be passed to \code{GetSignInteractions} (not used).}

\item{threshold}{A numeric with the FDR cut-off threshold used to take a
subset of significant interactions If \code{threshold=NULL}
then  all the interactions are returned.}

\item{ReturnedAs}{What type of object to return. Possible values are:
\describe{
\item{\code{'GInteractions'}}{ Will return a \code{\link[InteractionSet:GInteractions-class]{GInteractions}} object.
Each line corresponds to an interaction between two anchors (peaks). It provides information about the intervals of the
peaks, the anchor/peak summits, the interaction p-value/FDR and order (see \code{\linkS4class{GenomeMap}} for further details).}
\item{\code{'csv'}}{ Will save a .csv file in the \code{SaveTo}. The file will
be named after the \code{Prefix} parameter \code{Prefix_InteractionResults.csv}}
}}

\item{SaveTo}{The directory to save the csv file in case of \code{ReturnedAs=='csv'}.}

\item{Prefix}{A string which will be used a prefix on the object saved in \code{SaveTo}.}
}
\value{
For \code{\linkS4class{GenomeMap}} class, different outputs based on the \code{ReturnedAs} parameter.
}
\description{
\code{GetSignInteractions} subsets the significant interactions of a \code{\linkS4class{GenomeMap}} object
given a user specified FDR cut-off. It returns/saves different kind of objects based on the \code{ReturnedAs} value.
}
\details{
\code{\link{MACPETUlt}} at Stage 4 will run the interaction analysis for the ChIA-PET data.
This stage will proccess all the interactions in the data. After Stage 4 is run, the user can use the
\code{GetSignInteractions} function on the saved object of class  \code{\linkS4class{GenomeMap}},
 to subset the significant interactions given a FDR cut-off.
}
\examples{

#load Interaction data: (class=GenomeMap)
load(system.file('extdata', 'MACPET_GenomeMapData.rda', package = 'MACPET'))
class(MACPET_GenomeMapData)
GetSignInteractions(object=MACPET_GenomeMapData,
                    threshold = NULL,
                    ReturnedAs='GInteractions')

}
\references{
Vardaxis I, Drabløs F, Rye M and Lindqvist BH (2018). \emph{MACPET: Complete pipeline for ChIA-PET}.
To be published.
}
\seealso{
\code{\linkS4class{GenomeMap}}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
