% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSortestPath-methods.R
\name{GetSortestPath}
\alias{GetSortestPath}
\alias{GetSortestPath.default}
\alias{GetSortestPath.GenomeMap}
\title{Find sortest path between pairs of peaks given a set of significant interactions.}
\usage{
GetSortestPath(object, ...)

\method{GetSortestPath}{default}(object, ...)

\method{GetSortestPath}{GenomeMap}(object, threshold = NULL, ChrFrom, ChrTo,
  SummitFrom, SummitTo, ...)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{GenomeMap}}.}

\item{...}{Further arguments to be passed to \code{GetSortestPath} (not used).}

\item{threshold}{A numeric with the FDR cut-off threshold used to take a
subset of significant interactions If \code{threshold=NULL}
then  all the interactions will be used}

\item{ChrFrom}{Character specifying the chromosome of the "From" peak, for example "chr1".}

\item{ChrTo}{Character specifying the chromosome of the "To" peak.}

\item{SummitFrom}{Numeric specifying the peak summit of the "From" peak.}

\item{SummitTo}{Numeric specifying the peak summit of the "To" peak.}
}
\value{
A two-element list with the first element named \code{LinearPathLength}
for the linear length of the path between \code{SummitFrom} and \code{SummitTo},
and the second element named \code{ThreeDPathLength} for the 3D length of the sortest path
between \code{SummitFrom} and \code{SummitTo}.
}
\description{
\code{GetSortestPath} function subsets the significant interactions of a \code{\linkS4class{GenomeMap}} object
given a user specified FDR cut-off. Based on the leftover interactions, it creates a network and finds the length of the
sortest path between two user-specified peaks. Currently it only finds the sortest paths between intra-chromosomal peaks.
Therefore, the peaks have to be on the same chromosome.
}
\examples{

#load Interaction data: (class=GenomeMap)
load(system.file('extdata', 'MACPET_GenomeMapData.rda', package = 'MACPET'))
class(MACPET_GenomeMapData)
GetSortestPath(object=MACPET_GenomeMapData,
                    threshold = NULL,
                    ChrFrom="chr1",
                    ChrTo="chr1",
                    SummitFrom=10000,
                    SummitTo=1000000)

}
\references{
Vardaxis I, Drabløs F, Rye M and Lindqvist BH (2018). \emph{MACPET: Complete pipeline for ChIA-PET}.
To be published.
}
\seealso{
\code{\linkS4class{GenomeMap}}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
