% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLData.R
\docType{data}
\name{MACPET_GenomeMapData.rda}
\alias{MACPET_GenomeMapData.rda}
\title{Genomic interactions from ChIA-PET data}
\format{rda object of \code{\linkS4class{GenomeMap}} class.}
\description{
Genomic Interactions data from
ESR1 ChIA-PET subset data on human MCF-7.
}
\details{
\code{MACPET_GenomeMapData} is produced by the \code{\link{MACPETUlt}}
function at Stage 4 and it contains all the genomic interactions between the peaks
in the data.
}
\seealso{
\code{\linkS4class{GenomeMap}}
}
\author{
\describe{
\item{Main data creators}{Yijun Ruan, GIS, 2012-05-24}
\item{MACPET_GenomeMapData.rda creator}{Ioannis Vardaxis,
\email{ioannis.vardaxis@ntnu.no}}
}
}
\keyword{data}
