% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLData.R
\docType{data}
\name{MACPET_pintraData.rda}
\alias{MACPET_pintraData.rda}
\title{Intra-chromosomal PETs from ChIA-PET data}
\format{rda object of \code{\linkS4class{PIntra}} class.}
\description{
Intra-chromosomal PETs data from
ESR1 ChIA-PET subset data on human MCF-7.
}
\details{
\code{MACPET_pintraData} is produced by the \code{\link{MACPETUlt}}
function at Stage 2 and it contains the Intra-chromosomal PETs of the sample data.
}
\seealso{
\code{\link{SampleChIAPETData.bam}}, \code{\linkS4class{PIntra}}
}
\author{
\describe{
\item{Main data creators}{Yijun Ruan, GIS, 2012-05-24}
\item{MACPET_pintraData.rda creator}{Ioannis Vardaxis,
 \email{ioannis.vardaxis@ntnu.no}}
}
}
\keyword{data}
