% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakCallerUlt.R
\name{PeakCallerUlt}
\alias{PeakCallerUlt}
\title{Complete Binding Site Analysis Function}
\usage{
PeakCallerUlt(DataDir = "", DataFile = NULL, AnalysisDir = "",
  GenomePkg = NULL, fileSelf = NULL, fileIntra = NULL, fileInter = NULL,
  PopImage = TRUE, BlackList = TRUE, fileSelfFit = NULL, method = "BH",
  Stages = c(0:1))
}
\arguments{
\item{DataDir}{A string with the directory of the BAM/SAM file (stage 0 parameter). In case \code{DataFile} is an object of
\code{\link[GenomicAlignments]{GAlignmentPairs}} class, \code{DataDir} will not be taken into account.}

\item{DataFile}{A string with the name of the BAM/SAM file. It should end
with .bam or .sam accordingly (stage 0 parameter). Alternatively, an object of \code{\link[GenomicAlignments]{GAlignmentPairs}}
class with the paired-end tags to be analyzed.}

\item{AnalysisDir}{A string with the directory where you want the output
of the algorithm to be saved (stage 0,1 parameter).}

\item{GenomePkg}{A string with the name of the genome as defined by the
\code{\link[BSgenome]{available.genomes}}. Note that the genome has to be
installed before running the algorithm (stage 0 parameter).}

\item{fileSelf}{A string for the name of the Self-ligated PETs produced
as output (stage 0), and taken as input (stage 1).}

\item{fileIntra}{A string for the name of the Intra-chromosomal PET sproduced
as output (stage 0 parameter).}

\item{fileInter}{A string  for the name of the Inter-chromosomal PETs produced
as output (stage 0 parameter).}

\item{PopImage}{TRUE or FALSE indicating whether you want to create a figure
for the Self-ligated borders/population or not (stage 0 parameter).}

\item{BlackList}{TRUE or FALSE depending on whether the user wants to remove
peaks which overlap with black listed regions before inference. Currently
only available for \code{hg="hg19"}, however the user can provide his own as a
data.frame with the following columns (stage 0 parameter):
\describe{
\item{\code{Chrom}}{A character vector with the chromosome names ("chr1" etc)}
\item{\code{Region.Start}}{An integer vector with the start of the
black listed regions.}
\item{\code{Region.End}}{An integer vector with the end of the black
listed regions.}
}}

\item{fileSelfFit}{A string for the name of the fitted Self-ligated PETs produced
as output (stage 1 parameter).}

\item{method}{Which method to use for adjusting for FDR for the candidate peaks found,
using the  \code{\link[stats]{p.adjust}} function (stage 1 parameter).}

\item{Stages}{A numeric (0 or 1) or a vector (c(0,1)), indicating which stage(s) the
algorithm should run. If the vector is given, both stages will be run at once.}
}
\value{
All outputs are saved at the \code{AnalysisDir}. The output depents of the stages run:
\describe{
\item{Stage 0:}{
\describe{
\item{\code{fileSelf: }}{An object of \code{\linkS4class{PSelf}} class with
the Self-ligated PETs. The name is specified by the \code{fileSelf} argument.
 This output will be further used in stage 1.}
\item{\code{fileIntra: }}{An object of \code{\linkS4class{PIntra}} class with
the Intra-chromosomal PETs. The name is specified by the \code{fileIntra}
argument.}
\item{\code{fileInter: }}{An object of \code{\linkS4class{PInter}} class with
the Inter-chromosomal PETs. The name is specified by the \code{fileInter}
argument.}
\item{\code{Image: }}{If \code{PopImage=TRUE}, an image for the elbow method of the
 Self-ligated population and its cut-off from the Intra-ligated population.}
}
}
\item{Stage 1:}{
\describe{
\item{\code{fileSelfFit: }}{An object of \code{\linkS4class{PSFit}} class with
the peak information. The name is specified by the \code{fileSelfFit}
argument.}
}
}
\item{Stage 0:1:}{ All the above outputs.}
}
The function also creates a file named \code{MACPET_analysis.log}
with the progress of the analysis.
}
\description{
\code{PeakCallerUlt} is used for running a complete binding site
analysis of ChIA-PET data.
}
\details{
\code{PeakCallerUlt} runs a complete or partial binding site analysis
for ChIA-PET data, depending on the stages of the analysis the user wants to run.
The stages of the analysis are the following:
\describe{
\item{Stage 0:}{
PET classification stage. This stage takes the BAM/SAM ChIA-PET file, or an object of \code{\link[GenomicAlignments]{GAlignmentPairs}} class  as input
and it seperates the PETs into three categories:
Inter-chromosomal PETs (which connect two different chromosomes),
Intra-chromosomal PETs (which connect regions of the same chromosome) and
Self-ligated PETs (which are used for binding site analysis).
Self-ligated PETs are used for finding the protein binding sites (peaks),
while Intra- and Inter-chromosomal are used for interactions between
the peaks. Furthermore, it removes identically mapped PETs for reducing noise created by
amplification procedures. The algorithm uses the elbow-method to seperate the
Self-ligated from the Intra-chromosomal population.  Note that loading the data
into R might take a while depending on the size of the data.

}
\item{Stage 1:}{ Peak calling stage. This stage uses the Self-ligated PETs and it runs the
EM algorithm to find clusters which represent candidate peaks/binding sites in
2 dimentional space using skewed generalized students-t distributions (SGT).
 After the peak-calling analysis is done, the algorithm assesses the
significance of the candidate peaks using a local Poisson model.}
}
}
\section{Parallel}{

The function can be run in parallel using the
\code{\link[BiocParallel]{register}}
function. The user has to register a parallel backhead before starting the
function. However the algorithm is mainly written in C++ and it is therefore fast,
therefore no need of parallel
backhead is necessary.
}

\examples{
#Create a test forder on the desktop, or anywhere you want:
AnalysisDir=file.path(path.expand('~'),'Desktop')
dir.create(file.path(AnalysisDir,"MACPET.test"))
AnalysisDir=file.path(AnalysisDir,"MACPET.test")#where you will save
#the results

#load sample data to use in the algorithm and give inputs:
DataDir=system.file("extdata", package = "MACPET") #data path
DataFile="SampleChIAPETData.bam" #data name
PopImage=TRUE #Sample data, not very good classifiction results.
GenomePkg="BSgenome.Hsapiens.UCSC.hg19" #genome of the data.
fileSelf="pselfData" #name for Self-ligated
fileIntra="pintraData" #name for Intra-chromosomal
fileInter="pinterData" #name for Inter-chromosomal
BlackList=TRUE #remove PETs in black listed regions
Stages=c(0)#run single stage for the example.

#parallel backhead can be created using the BiocParallel package
#parallel backhead can be created using the BiocParallel package
#requireNamespace("BiocParallel")
#snow <- BiocParallel::SnowParam(workers = 1, type = "SOCK", progressbar=FALSE)
#BiocParallel::register(snow, default=TRUE)

#-run for the whole binding site analysis:
PeakCallerUlt(DataDir=DataDir,
              DataFile=DataFile,
              AnalysisDir=AnalysisDir,
              GenomePkg=GenomePkg,
              fileSelf=fileSelf,
              fileIntra=fileIntra,
              fileInter=fileInter,
              PopImage=PopImage,
              BlackList=BlackList,
              Stages=Stages)



#load results:
load(file.path(AnalysisDir,fileSelf))
class(pselfData) # see methods for this class

load(file.path(AnalysisDir,fileIntra))
class(pintraData) # see methods for this class

load(file.path(AnalysisDir,fileInter))
class(pinterData) # see methods for this class

#-----delete test directory:
unlink(AnalysisDir,recursive=TRUE)
}
\references{
\insertRef{macpetcite}{MACPET}

\insertRef{ENCODE_1}{MACPET}
}
\seealso{
\linkS4class{PSelf}, \linkS4class{PIntra},
\linkS4class{PInter},\code{\link{summary}},
\code{\link{AnalysisStatistics}}, \code{\link{plot}}
\code{\link{BiocParallel}},\code{\link{ConvertToPSelf}},
\code{\link{exportPeaks}},\code{\link{TagsToGInteractions}},
 \code{\link{PeaksToGRanges}},  \code{\link{PeaksToNarrowPeak}}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
