% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeaksToGRanges-methods.R
\name{PeaksToGRanges}
\alias{PeaksToGRanges}
\alias{PeaksToGRanges.default}
\alias{PeaksToGRanges.PSFit}
\title{Convert peaks to GRanges object}
\usage{
PeaksToGRanges(object, ...)

\method{PeaksToGRanges}{default}(object, ...)

\method{PeaksToGRanges}{PSFit}(object, threshold = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{PSFit}}.}

\item{...}{Further arguments to be passed to \code{PeaksToGRanges} (not used).}

\item{threshold}{A numeric with the FDR cut-off threshold used to take a
subset of significant peaks. If \code{threshold=NULL}
then  all the peaks are returned.}
}
\value{
For \code{\linkS4class{PSFit}} class, a
\code{\linkS4class{GRanges}}
object created by the estimated
peak information including metadata columns for the total PETs,
the p-value and the FDR of each peak.
}
\description{
\code{PeaksToGRanges} converts peaks of an object of \code{\linkS4class{PSFit}}
class to \code{\linkS4class{GRanges}} object.
}
\details{
\code{PeaksToGRanges} converts peak information into a
\code{\linkS4class{GRanges}} object. Each row in the
\code{\linkS4class{GRanges}} object represents a peak with 'CIQ.Up.start' and 'CIQ.Down.end' as start
and end coordinates, respectively (see \code{\linkS4class{PSFit}})
 Metadata will also include information for the total PETs,
the p-value and the FDR of each peak.
}
\examples{
#load Self-ligated data: (class=PSFit)
load(system.file('extdata', 'MACPET_psfitData.rda', package = 'MACPET'))
class(MACPET_psfitData)
PeaksToGRanges(object=MACPET_psfitData,threshold=1e-5)


}
\references{
Vardaxis I, Drabløs F, Rye M and Lindqvist BH (2018). \emph{MACPET: Model-based Analysis for ChIA-PET}.
To be published.
}
\seealso{
\code{\linkS4class{PSFit}}, \code{\link{PeaksToNarrowPeak}}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
