% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PInter-class}
\alias{PInter-class}
\alias{PInter}
\title{PInter S4 Class}
\description{
\code{\linkS4class{PInter}} class in a S4 class which
 inherits from the
\code{\link[InteractionSet]{GInteractions}} class and it contains
Inter-chromosomal data.
}
\details{
\code{\linkS4class{PInter}} class is created by the
 \code{\link{PeakCallerUlt}}
function at Stage 0.
}
\references{
\insertRef{macpetcite}{MACPET}
}
\seealso{
\code{\link{AnalysisStatistics}}, \code{\link{plot}},
\code{\link{summary}},
\code{\link{PeakCallerUlt}}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
