% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLData.R
\docType{data}
\name{pinterData.rda}
\alias{pinterData.rda}
\title{Inter-chromosomal PETs from ChIA-PET data}
\format{rda object of \code{\linkS4class{PInter}} class.}
\description{
Inter-chromosomal PETs data from
ESR1 ChIA-PET subset data on human MCF-7.
}
\details{
\code{pinterData} is produced by the \code{\link{PeakCallerUlt}}
function at Stage=0 and it contains the Inter-chromosomal PETs of the sample data.
}
\seealso{
\code{\link{SampleChIAPETData.bam}}, \code{\linkS4class{PInter}}
}
\author{
\describe{
\item{Main data creators}{Yijun Ruan, GIS, 2012-05-24}
\item{pinterData.rda creator}{Ioannis Vardaxis,
\email{ioannis.vardaxis@ntnu.no}}
}
}
\keyword{data}
