% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PIntra-class}
\alias{PIntra-class}
\alias{PIntra}
\title{PIntra Class}
\description{
\code{\linkS4class{PIntra}} class in a S4 class which
inherits from the
\code{\link[InteractionSet]{GInteractions}} class and it contains
Intra-chromosomal data.
}
\details{
\code{\linkS4class{PIntra}} class is created by the
\code{\link{PeakCallerUlt}}
function at Stage 0.
}
\references{
Vardaxis I, Drabløs F, Rye M and Lindqvist BH (2018). \emph{Model-based Analysis for ChIA-PET (MACPET)}.
To be published.
}
\seealso{
\code{\link{AnalysisStatistics}}, \code{\link{plot}},
\code{\link{summary}},
\code{\link{PeakCallerUlt}}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
