% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot}
\alias{plot}
\alias{plot.PInter}
\alias{plot.PIntra}
\alias{plot.PSelf}
\alias{plot.PSFit}
\title{plot methods for MACPET classes}
\usage{
\method{plot}{PInter}(x, ...)

\method{plot}{PIntra}(x, ...)

\method{plot}{PSelf}(x, ...)

\method{plot}{PSFit}(x, kind, RegIndex = NULL, threshold = NULL, ...)
}
\arguments{
\item{x}{An object of correct class used to create different plots.}

\item{...}{further arguments to be passed in the plot functions.}

\item{kind}{A string with one of the following arguments. Note that if a region visualization is plotted, the vertical lines represent peak-summits.
 \describe{
  \item{\code{PETcounts}}{ For a bar-plot of the PET-counts in each chromosome.}
  \item{\code{RegionCounts}}{ For a bar-plot for the region counts in each chromosome.}
  \item{\code{PeakCounts}}{ For a bar-plot for the Peak-counts in each chromosome.}
  \item{\code{RegionPETs}}{ For a ggplot for a visualization of the PETs in a region.}
  \item{\code{RegionTags}}{ For a ggplot for a visualization of the tags in a region. The tags are classified by stream (upper/lower)}
  \item{\code{PeakPETs}}{ For a ggplot for a visualization of the PETs in a region. The PETs are classified by the peak they belong to.}
  \item{\code{PeakTags}}{ For a ggplot for a visualization of the tags in a region. The tags are classified by the peak they belong to.}
  \item{\code{SigPETCounts}}{ For a bar-plot with the significant PET-counts in each chromosome.}
  \item{\code{SigRegionCounts}}{For a bar-plot with the significant region-counts in each chromosome.}
  \item{\code{SigPeakCounts}}{For a bar-plot with the significant peak-counts in each chromosome.}

 }}

\item{RegIndex}{an integer indicating which region to plot (1 means the biggest
in terms of total PETs.)}

\item{threshold}{The FDR cut-off when plotting the total significant peaks for each chromosome in the data.}
}
\value{
For the \code{\linkS4class{PInter}} class:
A network plot.
 Each node is a chromosome with size proportional to the total PETs of the
 corresponding chromosome. Edges connect chromosomes which have common PETs,
  where the thickness of an edge is proportional on the total number of PETs
  connecting the two chromosomes.

For the \code{\linkS4class{PIntra}} class:
A bar-plot. Each bar
represents the total number of Intra-chromosomal PETs for each chromosome in
the data.

For the \code{\linkS4class{PSelf}} class:
A bar-plot. Each bar
represents the total number of Self-ligated PETs for each chromosome in
the data.

For the \code{\linkS4class{PSFit}} class:
 Different plots depenting on the \code{kind} argument.
}
\description{
Different plot methods for the classes in the
\code{\link{MACPET}} package.
}
\examples{
#load Inter-chromosomal data:
load(system.file('extdata', 'MACPET_pinterData.rda', package = 'MACPET'))
class(MACPET_pinterData)
requireNamespace('igraph')
plot(MACPET_pinterData)
#load Intra-chromosomal data:
load(system.file('extdata', 'MACPET_pintraData.rda', package = 'MACPET'))
class(MACPET_pintraData)
requireNamespace('ggplot2')
plot(MACPET_pintraData)
#load Self-ligated data:
load(system.file('extdata', 'MACPET_pselfData.rda', package = 'MACPET'))
class(MACPET_pselfData)
requireNamespace('ggplot2')
plot(MACPET_pselfData)
#load Self-ligated data:
load(system.file('extdata', 'MACPET_psfitData.rda', package = 'MACPET'))
class(MACPET_psfitData)
requireNamespace('ggplot2')
plot(MACPET_psfitData,kind='PETcounts')
plot(MACPET_psfitData,kind='PeakCounts')
plot(MACPET_psfitData,kind='PeakPETs',RegIndex=1)
plot(MACPET_psfitData,kind='PeakTags',RegIndex=1)
}
\references{
Vardaxis I, Drabløs F, Rye M and Lindqvist BH (2018). \emph{MACPET: Model-based Analysis for ChIA-PET}.
To be published.
}
\seealso{
\code{\linkS4class{PSelf}},
\code{\linkS4class{PSFit}}
\code{\linkS4class{PInter}},\code{\linkS4class{PIntra}}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
