% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PSelf-class}
\alias{PSelf-class}
\alias{PSelf}
\title{PSelf S4 Class}
\description{
\code{\linkS4class{PSelf}} class in a S4 class which
 inherits from the
\code{\link[InteractionSet:GInteractions]{GInteractions}} class and it contains Self-ligated
PETs from ChIA-PET experiment. Furthermore it also contains the following in the
\code{\link[S4Vectors:metadata]{metadata}} field:
 \describe{
 \item{\code{Self_info}}{ A data.frame with the count statistics for the total PETs in each chromosome.}
 \item{\code{SLmean}}{The mean size of the PETs in the data.}
 \item{\code{MaxSize}}{Maximum size of self-ligated PETs.}
 \item{\code{MinSize}}{Minimum size of self-ligated PETs.}
 }
}
\details{
\code{\linkS4class{PSelf}} class is created by
the \code{\link{MACPETUlt}}
function at Stage 2.
}
\references{
Vardaxis I, Drabløs F, Rye M and Lindqvist BH (2018). \emph{MACPET: Model-based Analysis for ChIA-PET}.
To be published.
}
\seealso{
\code{\link{AnalysisStatistics}}, \code{\link{plot}},
\code{\link{summary}},
\code{\link{MACPETUlt}}, \code{\link{ConvertToPSelf}}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
