% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLData.R
\docType{data}
\name{pselfData.rda}
\alias{pselfData.rda}
\title{Self-ligated PETs from ChIA-PET data}
\format{rda object of \code{\linkS4class{PSelf}} class.}
\description{
Self-ligated PETs data from
ESR1 ChIA-PET subset data on human MCF-7.
}
\details{
\code{pselfData} is produced by the \code{\link{PeakCallerUlt}}
function at Stage 0 and it contains the Self-ligated PETs of the sample data.
}
\seealso{
\code{\link{SampleChIAPETData.bam}}, \code{\linkS4class{PSelf}}
}
\author{
\describe{
\item{Main data creators}{Yijun Ruan, GIS, 2012-05-24}
\item{pselfData.rda creator}{Ioannis Vardaxis,
\email{ioannis.vardaxis@ntnu.no}}
}
}
\keyword{data}
