% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLData.R
\docType{data}
\name{psfitData.rda}
\alias{psfitData.rda}
\title{Self-ligated PETs from ChIA-PET data}
\format{rda object of \code{\linkS4class{PSFit}} class.}
\description{
Self-ligated PETs data from
ESR1 ChIA-PET subset data on human MCF-7.
}
\details{
\code{psfitData} is produced by the \code{\link{PeakCallerUlt}}
function at Stage 1 and it contains the self-ligated PETs of the sample data after
calling for candidate peaks.
}
\seealso{
\code{\link{SampleChIAPETData.bam}},
\code{\linkS4class{PSFit}}
}
\author{
\describe{
\item{Main data creators}{Yijun Ruan, GIS, 2012-05-24}
\item{psfitData.rda creator}{Ioannis Vardaxis,
\email{ioannis.vardaxis@ntnu.no}}
}
}
\keyword{data}
