% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLData.R
\docType{data}
\name{SampleChIAPETData.bam}
\alias{SampleChIAPETData.bam}
\title{Subset sample of ChIA-PET data}
\format{A BAM file.}
\source{
\url{https://www.encodeproject.org/experiments/ENCSR000BZZ/}
}
\description{
A subset of ChIA-PET data:
\describe{
\item{Target: }{ESR1}
\item{Biosample summary: }{Homo sapiens MCF-7}
\item{GEO: }{GSM970212}
}
}
\references{
\insertRef{ENCODE_1}{MACPET}
}
\seealso{
\code{\link{pselfData.rda}}, \code{\link{pintraData.rda}},
\code{\link{pinterData.rda}},
\code{\link{psfitData.rda}}
}
\author{
Yijun Ruan, GIS, 2012-05-24 (main data creators)
}
\keyword{data}
