% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLData.R
\docType{data}
\name{SampleChIAPETDataRead_2.bam}
\alias{SampleChIAPETDataRead_2.bam}
\title{Second reads from a subset of ChIA-PET data in \code{\link{SampleChIAPETData.bam}}}
\format{A BAM file.}
\source{
\url{https://www.encodeproject.org/experiments/ENCSR000BZZ/}
}
\description{
Second reads from a subset of ChIA-PET data in \code{\link{SampleChIAPETData.bam}}:
\describe{
\item{Target: }{ESR1}
\item{Biosample summary: }{Homo sapiens MCF-7}
\item{GEO: }{GSM970212}
}
}
\references{
Consortium EP (2012) \emph{An integrated encyclopedia of DNA elements in the human genome.}.
 Nature, 489(7414), pp. 57–74. \url{http://dx.doi.org/10.1038/nature11247}.
}
\seealso{
\code{\link{SampleChIAPETData.bam}},  \code{\link{ConvertToPE_BAM}}
}
\author{
Yijun Ruan, GIS, 2012-05-24 (main data creators)
}
\keyword{data}
