% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagsToGInteractions-methods.R
\name{TagsToGInteractions}
\alias{TagsToGInteractions}
\alias{TagsToGInteractions.default}
\alias{TagsToGInteractions.PSFit}
\title{Convert PETs to GInteractions object}
\usage{
TagsToGInteractions(object, ...)

\method{TagsToGInteractions}{default}(object, ...)

\method{TagsToGInteractions}{PSFit}(object, threshold = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{PSFit}}.}

\item{...}{(not used).}

\item{threshold}{A numeric for the FDR threshold used to take a subset of
significant peaks/binding sites. If \code{threshold=NULL} then
all the peaks are returned.}
}
\value{
For \code{\linkS4class{PSFit}} class:
A \code{\link[InteractionSet:GInteractions-class]{GInteractions}} object containing PETs from all
 the peaks found in the data (removing noisy and insignificant PETs).
 Furthermore, it also includes information about
 the binding sites which can be accessed via the
 \code{\link[S4Vectors:metadata]{metadata}} function.
}
\description{
\code{TagsToGInteractions} converts the PETs of an object of \code{\linkS4class{PSFit}}
class to \code{\link[InteractionSet:GInteractions-class]{GInteractions}} object.
}
\examples{
#load Self-ligated data: (class=PSFit)
load(system.file('extdata', 'MACPET_psfitData.rda', package = 'MACPET'))
class(MACPET_psfitData)
object=TagsToGInteractions(object=MACPET_psfitData,threshold=1e-5)
object
S4Vectors::metadata(object)$Peaks.Info #peak/binding site information

}
\references{
Vardaxis I, Drabløs F, Rye M and Lindqvist BH (2018). \emph{MACPET: Model-based Analysis for ChIA-PET}.
To be published.
}
\seealso{
\code{\linkS4class{PSFit}}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
