% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertToPSelf-methods.R
\name{ConvertToPSelf}
\alias{ConvertToPSelf}
\alias{ConvertToPSelf.default}
\alias{ConvertToPSelf.GInteractions}
\title{Convert GInteraction object to PSelf object}
\usage{
ConvertToPSelf(object, ...)

\method{ConvertToPSelf}{default}(object, ...)

\method{ConvertToPSelf}{GInteractions}(object, GenomePkg, BlackList, ...)
}
\arguments{
\item{object}{An object of \code{\link[InteractionSet]{GInteractions}} class.}

\item{...}{not used.}

\item{GenomePkg}{See  \code{\link{PeakCallerUlt}}.}

\item{BlackList}{See  \code{\link{PeakCallerUlt}}.}
}
\value{
An object of class \code{\linkS4class{PSelf}}.
}
\description{
\code{ConvertToPSelf} converts a GInteractions object to
class to \code{\linkS4class{PSelf}} object.
}
\details{
\code{\link{PeakCallerUlt}} with Stage=0 is the first function the user can
use for separating Inter-chromosomal, Intra-chromosomal and Self-ligated PETs by giving
the whole paired-end BAM/SAM file as input. However the user might only have
Self-ligated data available and already separated from the Inter/Intra-chromosomal
PETs. \code{ConvertToPSelf} can then be used in the Self-ligated data to convert
a \code{\link[InteractionSet]{GInteractions}} object containing only the Self-ligated
PETs to a \code{\linkS4class{PSelf}} class for further analysis.
}
\examples{
#load Self-ligated data: (class=PSelf)
load(system.file("extdata", "pselfData.rda", package = "MACPET"))
class(pselfData)

object=pselfData
#--remove information and convert to GInteractions:
S4Vectors::metadata(object)=list(NULL)
class(object)="GInteractions"
GenomePkg="BSgenome.Hsapiens.UCSC.hg19" #genome of the data.
BlackList=TRUE

object=ConvertToPSelf(object=object,GenomePkg=GenomePkg,BlackList=BlackList)
class(object)

}
\references{
Vardaxis I, Drabløs F, Rye M and Lindqvist BH (2018). \emph{Model-based Analysis for ChIA-PET (MACPET)}.
To be published.
}
\seealso{
\code{\linkS4class{PSelf}}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
