% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PInter-class}
\alias{PInter-class}
\alias{PInter}
\title{PInter S4 Class}
\description{
\code{\linkS4class{PInter}} class in a S4 class which
 inherits from the
\code{\link[InteractionSet:GInteractions-class]{GInteractions}} class and it contains
Inter-chromosomal data.
}
\details{
\code{\linkS4class{PInter}} class is created by the
 \code{\link{MACPETUlt}}
function at Stage 2.
}
\references{
Vardaxis I, Drabløs F, Rye M and Lindqvist BH (2018). \emph{MACPET: Model-based Analysis for ChIA-PET}.
To be published.
}
\seealso{
\code{\link{AnalysisStatistics}}, \code{\link{plot}},
\code{\link{summary}},
\code{\link{MACPETUlt}}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
