% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PSelf-class}
\alias{PSelf-class}
\alias{PSelf}
\title{PSelf S4 Class}
\description{
\code{\linkS4class{PSelf}} class in a S4 class which
 inherits from the
\code{\link[InteractionSet]{GInteractions}} class and it contains Self-ligated
PETs from ChIA-PET experiment. Furthermore it also contains the following in the
\code{\link[S4Vectors]{metadata}} field:
 \describe{
 \item{\code{Self_info}}{ A data.frame with the count statistics for the total PETs in each chromosome.}
 \item{\code{SLmean}}{The mean size of the PETs in the data.}
 \item{\code{GenInfo}}{Information about the genome.}
 \item{\code{MaxSize}}{Maximum size of self-ligated PETs.}
 \item{\code{MinSize}}{Minimum size of self-ligated PETs.}
 }
}
\details{
\code{\linkS4class{PSelf}} class is created by
the \code{\link{PeakCallerUlt}}
function at Stage=0.
}
\references{
\insertRef{macpetcite}{MACPET}
}
\seealso{
\code{\link{AnalysisStatistics}}, \code{\link{plot}},
\code{\link{summary}},
\code{\link{PeakCallerUlt}}
}
\author{
Ioannis Vardaxis, \email{ioannis.vardaxis@ntnu.no}
}
