% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLData.R
\docType{data}
\name{SampleChIAPETData.bam}
\alias{SampleChIAPETData.bam}
\title{Subset sample of ChIA-PET data}
\format{A BAM file.}
\source{
\url{https://www.encodeproject.org/experiments/ENCSR000BZZ/}
}
\description{
A subset of ChIA-PET data:
\describe{
\item{Target: }{ESR1}
\item{Biosample summary: }{Homo sapiens MCF-7}
\item{GEO: }{GSM970212}
}
}
\references{
Consortium EP (2012) \emph{An integrated encyclopedia of DNA elements in the human genome.}.
 Nature, 489(7414), pp. 57–74. \url{http://dx.doi.org/10.1038/nature11247}.
}
\seealso{
\code{\link{MACPET_pinterData.rda}}, \code{\link{MACPET_pintraData.rda}},
\code{\link{MACPET_pselfData.rda}},
\code{\link{MACPET_psfitData.rda}}, \code{\link{SampleChIAPETDataRead_1.bam}},
 \code{\link{SampleChIAPETDataRead_2.bam}}
}
\author{
Yijun Ruan, GIS, 2012-05-24 (main data creators)
}
\keyword{data}
